[sustainability] add a few static pages, remove unrelated routes, and add tests (#1173)
This disables all non-api routes which aren't explicitly marked for sustainability site. At the moment, the exact same templates are used for feedback and disclaimers, but I imagine we would want to add specialized text for the rest (or link to main site).

Explicitly disabling the others so we won't have URL mixups with the various subdomains. Also ensures we thoughtfully add each page back. Added all API routes, though, to simplify adding tools in the future.

Added basic tests to ensure the pages render without error in each environment. We'd want to add link crawlers / checkers for external links, if we add more links to main site.