[topic] Proto update - move stat_var_metadata to the category layer (#1441)
I found that svm definitions were duplicated multiple times across blocks, and specifying overrides for some tiles were cumbersome. This also removes some ambiguity as to which svm is used for tiles like highlight and map which only use the first svm. Eventually, we can centralize all data fetching for a category.

Updated tests and configs.

Also moved all typed interfaces for the config proto to a single file that mirrors the proto.