Refactor map value data fetch, compute and time slider logic (#1797)
* Time Slider: simplify logic by removing the callback functions and let
props.current_date to trigger its rendering. The component is
responsible for set date context(). With this, data is fetched for each
date as if a normal rendering in map tool. This can save the long
initial data load, but could result in longer frame rate.
* Make compute hooks for map values and breadcrumb values, sample dates
and other chart data.
* Date context: Separate date as its own context so it can be set in
time slider with minimal effect in render logic.
* Loading spinner should be re-visited.