[private dc] Use a config for tool landing page links (#1818)
Create json objects that is passed into the template. The component uses
this to render the landing page links. This can be customized for the
private dc instances.

Using JSON seemed simplest for configuration purposes, and will be more
inline with how other templates are overwritten for private dc (I tried
using protos which had a lot more config / creation overhead).

If this looks good, will apply this to other tools.