Allow dynamic host configuration for website endpoints (#1857)
Currently the website backend assumes that Mixer runs on localhost. In
order to decouple mixer from website at the host level and deploy them
independently with helm, we need to break this assumption.

This change allows the mixer host to be configured using ingress rules.
Ingress rules are a mapping of host:port -> list of url paths that the
host:port serves.

On a high level, the source of truth is the global variable "endpoints",
which is configured during server startup time. Configuring means
matching the ingress path (which may contain wildcard) with paths in
flask server.

TESTED=unit test

Co-authored-by: Alex Chen <alexyfchen@google.com>