[disaster] Flask changes for updating topic page proto (#1874)
For the new disaster dashboard, we will be making it config driven &
doing this by utilizing what we had for the topic pages. We would also
like to make the config style page something that can be re-used in the
future by other dashboards/apps. To differentiate between the topic page
app and the shared config style pages, we will name shared
proto/components/etc SubjectPage.

To make changes easier to view when reviewing, I'm splitting up the
change into multiple PRs. This PR has:
- proto changes: naming change + change to the structure of Tile and
Block
- updates to topic page configs, flask code & React code to be
compatible with the new proto & allow current topic pages to render the
same as on prod