[NL Interface] No Place detection: context and default (#2001)
Up until now, we always use a default place if a place was not detected
in the query. Default was USA. But now, we could have previous context
to use. So here is the logic:

0. If place detected, use it. Always.
1. If no place detected and no context: use default (USA)
2. If no place detected and context: use from the most recent context.
If there is none in the recent context, go back to the default place.

Tested this locally on a series of queries and it seems to do the right
thing.

Note: eventually, we must find a way to do this in the same place we are
inheriting SVs from context. Problem right now is that it will be a big
change because currently the DataSpec is loaded with SVs expecting there
to be a place.