[NL Interface] Refactoring + Query parsing enhancement: remove stop words (#2042)
This is a fairly major enhancement. Today, we only remove the "place"
detected (rather the part of the query which has the words for the
place) from the query before doing SV detection. This means the query
still has lots of stop words ("a", "and", "what about", "cities",
"counties", "most", "best" etc). This has a major drawback for the
embeddings matches: in the embeddings space, the semantic meaning can
often get tripped up these extra words.

As an example, in the current mechanism (without stop words removal),
the following string:

"Which counties have agriculture workers in California" matches to the
following top SVG/SV:

dc/g/Farm_FarmType-Cropland
Area_Farm_Cropland

Try on https://dev.datacommons.org/nl/

However, "Which counties have the most agriculture workers in
California" matches to dc/hlxvn1t8b9bhh which is what we always want in
this case.

Asking for agriculture workers with the word "**counties**" makes things
match closer to the "*_Cropland" SVs which is not what one would expect.
But "Which **cities** have agriculture workers in California" matches to
dc/hlxvn1t8b9bhh. It appears that in the semantic space, including
"counties" appears to bias the meaning towards "area" and "cropland".

With the removal of these stop and classification detection words, we
can do a lot better. All of the queries above now bring up
dc/hlxvn1t8b9bhh as the top match.

I have also tested all Scenarios in the demo script and they both work
100% the same as the case without the stopwords removal.