Refactor Flask config loading (#2295)
Currently there are two issues for config class:
1) They all resides in one file, makes it harder to track
2) Local copy of a custom config needs to set "ENV_NAME" for them to be
grouped together.

This refactor tries to setup a mini-framework, such that "FLASK_ENV"
directly maps to a module (with the same name), that has "Config" and
"LocalConfig" class.

This makes it much easier to add a custom DC config. In this PR, a top
level "custom_dc" folder is created to hold a "custom.py" template
config module which is readily used. This folder will also contain other
static files.