[Custom DC] Various setup improvements (#2349)
Cluster
- Migrate the cluster management from shell script to Terraform module.
This allows Terraform to also delete the GKE cluster, 1 step closer for
E2E test
- Allow zonal cluster instead of enforcing regional clusters, which
should cut the resource by 2/3. Location specifies whether a zonal or a
regional cluster is created. ex: "us-cental1" -> regional,
"us-central1-a" -> zonal. The difference is 99.95% SLA vs 99.5% SLA. I
think most POCs/dev projects makes sense to be zonal.

Helm chart
- Allow gcr repo project to be specified, this is a follow up for
@shifucun's support for building images in custom dc project
- Move helm out of Terraform management. In the past few weeks, issue 1
is that it was easy to get into a circular situation where the state
needs to be read at the beginning (reading what resource a cluster has
requires kube config), but the kube config is yet to be fetched because
it happens in the part after reading the states. May be cleaner to
separate GCP resource level and k8s resource levels.
- This does not affect GCP level resource lifecycle, as even if pods and
services are in a cluster, the cluster can still be deleted.
- Taking the k8s resources out of the Terraform lifecycle may make
things easier when integrating with Argo, as Argo changes will not touch
Terraform at all.

---------

Co-authored-by: Alex Chen <alexyfchen@google.com>