[NL Interface] Updating the embeddings generation script (v2) (#2555)
The script does the following:

1. Downloads the latest SV file and writes under the sheets/ folder
(note: at this point, the file being used is different from the one
being used in the live server to generate embeddings--once this new
process is complete, the old Google Sheets file will be retired).
2. Uses other sources of alternate sentences/descriptions based on PaLM
or other sources.
3. Build the embeddings using the alternatives in the source file (1)
and the alternatives (2).
4. Writes the latest CSV files for the concatenated final
descriptions/alternatives to the repo + pushes the generated embeddings
to GCS.

This was tested by comparing the generated embeddings file size with the
existing embeddings files. The file size should be larger now (because
of the PaLM alternatives). I also tested without including PaLM
alternatives and the file size was exactly equal to the existing latest
embeddings file. Also checked that all alternate strings in the existing
embeddings file were also available in the updated one.

Note: the script (build_embeddings_v2.py) is not currently the default
script. It can be used (and migrated to build_embeddings.py) once this
PR is approved and all looks good.