Add a framework to get stable screenshot result (#2801)
A lot of charts, widgets are fetched async by using react effect and
state. For example on place page, there are tens of charts that show up
at different times. This makes screenshot unstable.

This PR adds a framework to wait for all the charts/widgets loading
before taking a screenshot. The idea is simple, add a class for the
holder element that are initially rendered and add a class for the
chart/widget element that are updated asynchronously. The tests first
gather all the holder elements and check when each holder element has
the desired element present.

We should use in other web driver tests as well.