[website revamp] add NL features needed for the homepage (#2877)
- add &d=1 flag which will disable the delay when autorunning a query on
NL Interface. This way, when user enters a query in the homepage search
and clicks submit, the result will be loaded without delay
- add &topic= flag which will show a list of example queries for a
topic. Note: these examples I've added are just to show what this
behavior looks like & needs to be updated to actual queries
- also fix click on datacommons in the nav bar to go back to old hompage
instead of the new one


![screen-recording-_98_](https://github.com/datacommonsorg/website/assets/69875368/6f838908-fcda-4e87-91c7-a180181b228f)

Co-authored-by: Prashanth R <shanth@google.com>