[NL Interface] Integration test failure fix (#2887)
This small fix fixes the broken integration test.

The current failure is happening because the Maps API has started to
return results in a different order where searching for "new york state"
is returning "Staten Island" above "New York State". In this fix, we
check the results returned and prefer an exact (lowercase) match instead
of the first result.