[NL Interface] Bug fix: check if the predownloaded model can successfully be loaded (#2898)
This bug was preventing the local nl_server from starting. 

Issue: on OS X any folders/files under the `/tmp/` folder are deleted if
unused for some time (ref thanks to @pradh :
https://apple.stackexchange.com/questions/438930/does-macos-delete-files-in-tmp-periodically).
But looks like this cleanup does not properly remove a parent folder if
there is an empty folder contained in it. This resulted in our
nl_server's initialization to think the model is already downloaded
(because we were only checking for path existence) but when loading the
model it threw an error.

Fix is to check at startup time if the model can actually be loaded
successfully. If not, remove the path entirely and then re-download the
model folder and files.