[Library] Allow specifying api root (#2967)
Adds `apiRoot` as an attribute to each chart component, which specifies
the host we wish to use for the webserver's apis (localhost:8080,
autopush.datacommons.org, datacommons.org, etc).

Mainly useful internally for development as we test components / use
components to debug new data in autopush. External users should never
need to set this.