Solve conflict state from topic, place and query and only rely on <topic,place> from url hash (#3024)
In a page with a combination of <topic, place>, user can issue a query
with totally different place and topic. In this case, we don't want to
store conflicting topic, place and query on the url. Instead, first call
"detect" API to retrieve the desired topic and place, update the url
hash (also erase the query). So the page would refresh based on the
detected topic and place.