Support >1 place for time-delta and add to answer places (#3282)
When we have answer places, we want to be able to add a few more places
from the query to it. Like US and Germany in the demo.

Previously, time-delta line charts were only supported for contained-in
scenario. Support that for the case where we have multiple places too.

Also:
1. Drop the custom Explore hacks (hoisting topics and having different
SV limits)
2. Reduce the number of Timeline chart places returned in time-delta

Enables:
https://screencast.googleplex.com/cast/NjQwOTMxNTUzNjYwMTA4OHxkN2I4OGVhOS03Nw