[NL Interface] Make all chart/block titles be title-case (#3316)
All chart and block titles are not title-case. Also updating all the
goldens (where all the updates are for the title casing).

Title-casing now does several things:

1. Upper-case a word if it does not start with a number, e.g. to avoid
"7th" becoming "7Th".
2. Upper-case a word if it isn't a connecting word like "a" or "an" etc.
This brings us to parity with what happens in prod. This includes words
like "the" and "and" etc.
3. Do not upper-case a word which is not a lower-case word, e.g. "GDP"
should not become "Gdp".
4. Only replace " - " with "-" if surrounded on both sides by a digit.