Implement superlatives and curtail llm fallback more (#3355)
1. Restructure the size_type handler into a more general superlative
handler. Also use it for `list` queries (for which the name is a slight
misnomer).

2. Be stricter about fallbacking back to LLM. Specifically, a query like
[which communities in Kansas as uninsured] is considered a multi-sv
query (because "communities" and "uninsured" as different vars separated
by Kansas). Have a much higher score threshold for the individual SV
parts. So really only when there are high-matching SVs do we consider it
multi-sv.


![image](https://github.com/datacommonsorg/website/assets/4375037/0e9f9a9c-77ae-46c2-9fc3-ddb3f921d3b8)


![image](https://github.com/datacommonsorg/website/assets/4375037/3cc0e351-576b-4ea5-9418-4d2430bafe71)