Update Place Overview UI (#3825)
Lays the groundwork for the UI update for the place overview of the
place pages based on figma mocks. This PR leaves the summary text as a
paragraph in the overview tile.

Left for follow-up PRs:
* Switching to sentence-by-sentence highlights will come in a followup
PR.
* Updating the other components on the place pages to match Figma mocks.

Pages With Summary
![Screenshot 2023-12-04 at 9 25
48 AM](https://github.com/datacommonsorg/website/assets/4034366/90e24d78-6b05-4587-9c11-e95a12702bb2)


Pages Without Summary
![Screenshot 2023-12-04 at 9 25
57 AM](https://github.com/datacommonsorg/website/assets/4034366/231b76ce-004a-428c-ba08-22802d6fc661)

Testing:
Verified changes locally don't break the place pages, and that the place
overview shown on the NL interface does not change.