Handle variable comparison in a single place (#3875)
When we find multi-sv query, and the place is not one with a default
child-type (like country / state), we currently fallback to a single-sv
result.

Instead, this PR extends `correlation` fulfiller to plot multiple SVs in
a single timeline/bar chart (by leveraging `simple` fulfiller),
**after** checking that the SVs are plottable in a single chart (i.e.,
are both PC compat/incompat and have same unit). See `is_coplottable`
function.

Additionally,

* Avoid spurious "Sorry, there were no relevant statistics for places in
Sunnyvale. See results for Sunnyvale." with the additional check in
`_maybe_set_fallback`.

* When there are 2 SVs the chart titles currently have "SV1 and more",
which looks odd for the variable comparison query. So improve that for
the 2 SV case.

* Unless the correlation-query returns a SCATTER, we were not returning
the proper topics. Fix that in `_get_plotted_orig_vars`.

TODO: "native born" is currently broken in the embeddings index because
it returns "US Native Born Outside US" (rather than a native that was
born either in US or not). That needs further fixing.


![image](https://github.com/datacommonsorg/website/assets/4375037/c7c312a7-5119-49c2-baff-885cdc84a47c)