[Place Page SEO] Check-in place page summary generation script (#3912)
Adds the scripts used to generate the summary text shown on our place
pages.

There are two scripts and two sets of variables and text provided:

(1) A "ranking based" summary text, where sentences are written for
variables a place ranks highly in. This is the approach currently used
for the summaries we have up. This includes the following files:
* `build_ranking_based_summaries.py`
* `stat_vars_detailed.json`
* `place_summary_content_us_states.json` 

(2) A templated summary, where we write sentences about the same set of
stat vars for every place. These summaries are intended to be used as
starting draft to feed into LLMs to rewrite into a final draft.
* `build_template_summaries.py`
* `stat_vars_to_highlight.json`
* `priority_places.json`
* `place_summaries_from_template.json`