Remove stat var hierarchy config from custom dc base.html (#3937)
A quick fix for a bug in Custom DC. The script in custom/base.html leads
to a server error broken page when following the custom DC quick start
guide.

This PR removes the script so that users can follow the custom DC quick
start guide. A more complete fix to re-enable features from #3911 will
come in a follow-up PR once we determine root cause of the issue.