Remove place page for West Berlin (nuts/DE301) from sitemap (#4107)
Removes the place page for West Berlin (nuts/DE301) from our sitemap.
This place page does not have any charts on it, and because this city no
longer exists, we won't be adding new charts to it. This PR removes it
from our sitemaps so it doesn't get indexed.

Also removes the unused `yapf` in requirements.txt which was causing
some version conflict errors.