Feature/home page react (#1)
* Add semicolon to server script

* Update .gitignore for IDE

* Moved the onload functionality into a separate function and fixed typing.

* Conversion of homepage.html to React. The homepage now passes template variables to main.ts, which initializes a React app. All homepage components are now implemented via this React app.

Typing files that could potentially be shared later were put into typing files in the shared directory.

* The backend is now sending the topics to the homepage as json, removing the need for json treatment on the front-end.

* The header and footer of the base template are now React apps.

A note that we will have to convert in a subsequent push all existing onLoad calls for existing React Apps to use an event listener to allow multiple apps to be called on one page from different places.

There is also room for some refactoring and cleanup.

* Converted the extraction of translated label meta-data into a loop rather than having it be explicit.

* Conversion of window.onload calls to window.addEventListener.

The trigger for this change is the chance that we might get multiple window.onload assignments in different places (with only the last applied assignment being run). This became an issue when the base template was given JavaScript (for the header and footer).

It isn't a further issue yet because those are using addEventListener, but by pre-emptively converting all window.onloads, we prevent future clashes.

* With the goal both of abstracting the data from the content of the header and footer (the menus) and to facilitate customization, the base components have been reworked.

The metadata passed to the front-end via the Jinja templates have been separated out from the main template, and the dictionaries they tie into are no longer strict TypeScript objects, but records. They store as keys the original text, and as a value the translation. The objects themselves are proxies that return the original text back if a translation does not exist. This makes the interaction between the templates, the TypeScript/JavaScript and the React components more flexible.

The routing has been reworked in the same way.

The header and footer menus themselves are json files located in a data directory inside the base app section. They can be overridden to make custom menus by saving them as "header_custom_dc.json" in the same directory and customizing that file (likewise with footer).

The webpack configuration is updated to facilitate these optional imports.

* The json that supplies the header and footer menu items is now moved to the backend and injected into the context of every endpoint. We still need to consider a simple way to allow for the overriding of this for custom templates.

* Update .gitignore for IDE

* Moved the onload functionality into a separate function and fixed typing.

* Conversion of homepage.html to React. The homepage now passes template variables to main.ts, which initializes a React app. All homepage components are now implemented via this React app.

Typing files that could potentially be shared later were put into typing files in the shared directory.

* The backend is now sending the topics to the homepage as json, removing the need for json treatment on the front-end.

* The header and footer of the base template are now React apps.

A note that we will have to convert in a subsequent push all existing onLoad calls for existing React Apps to use an event listener to allow multiple apps to be called on one page from different places.

There is also room for some refactoring and cleanup.

* Converted the extraction of translated label meta-data into a loop rather than having it be explicit.

* Conversion of window.onload calls to window.addEventListener.

The trigger for this change is the chance that we might get multiple window.onload assignments in different places (with only the last applied assignment being run). This became an issue when the base template was given JavaScript (for the header and footer).

It isn't a further issue yet because those are using addEventListener, but by pre-emptively converting all window.onloads, we prevent future clashes.

* With the goal both of abstracting the data from the content of the header and footer (the menus) and to facilitate customization, the base components have been reworked.

The metadata passed to the front-end via the Jinja templates have been separated out from the main template, and the dictionaries they tie into are no longer strict TypeScript objects, but records. They store as keys the original text, and as a value the translation. The objects themselves are proxies that return the original text back if a translation does not exist. This makes the interaction between the templates, the TypeScript/JavaScript and the React components more flexible.

The routing has been reworked in the same way.

The header and footer menus themselves are json files located in a data directory inside the base app section. They can be overridden to make custom menus by saving them as "header_custom_dc.json" in the same directory and customizing that file (likewise with footer).

The webpack configuration is updated to facilitate these optional imports.

* The json that supplies the header and footer menu items is now moved to the backend and injected into the context of every endpoint. We still need to consider a simple way to allow for the overriding of this for custom templates.

* Updated the home page to use the newer routing flow introduced during work on the base section.

* Update Ids set in loop to be set based on a sluggified version of the label.

---------

Co-authored-by: Jennifer Blumberg <jjblumberg@gmail.com>