WIP: A functional implementation of the rich menu. A note that this has not been styled to match the design.
As part of this commit, we have defined the header structure for the data source that populates the rich menu (a somewhat more complicated structure than the previous header).

We have separated the natural languages search bar component into a parent that calls one of two variant implementations in order to accommodate both the header version (which will be quite different in how it looks) and the standard version.