Home page, about page and build your own Data Commons page, and footer (#4599)
### Description

This PR implements the new home, about page and “build your own Data
Commons” page, as well as the footer.

The about page is located at `/about` and the “build your own Data
Commons” is located at `/build`.

As part of this PR, we have begun a process of refactoring some of the
functionality into components, including generic components that can be
used in other areas of the application later (such as text, text/media,
quote and carousel components).

### A Note

This PR is not fully finished yet. There is some work pending on the
carousel, as well some fine-tuning of some of the styling and some
refactoring. However, it is substantially ready for review.

### Description

This PR implements the new home, about page and “build your own Data
Commons” page, as well as the footer.

The about page is located at `/about` and the “build your own Data
Commons” is located at `/build`.

As part of this PR, we have begun a process of refactoring some of the
functionality into components, including generic components that can be
used in other areas of the application later (such as text, text/media,
quote and carousel components).

### A Note

This PR is not fully finished yet. There is some work pending on the
carousel, as well some fine-tuning of some of the styling and some
refactoring. However, it is substantially ready for review.

---------

Co-authored-by: Jennifer Blumberg <jjblumberg@gmail.com>
Co-authored-by: Pablo Noel <pablo@next-solutions.ca>