There is functionality associated with an id in the search box (which cannot yet be updated). In the long term, we could look at changing this so that it is no longer id based, but for now, there should only be one search box on the screen at the same time.
So rather than simply hiding the desktop search box in mobile view, and vice versa, we are now selectively rendering the header in the containing React component.

To facilitate this, this push also adds in a breakpoints hook that allows us to apply Bootstrap-like media queries on the TypeScript side (this hook may be more generally useful).